# 2/first-frontend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
        "/clienti": {
            "get": {
                "tags": [
                    "Clienti"
                ],
                "summary": "Returneaza toti clientii",
                "description": "Returneaza toti clientii",
                "operationId": "1649e86b74af2c06f91fd92bf6e168a5",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clienti"
                ],
                "summary": "Update clientul cu id-ul dat",
                "description": "Update clientul cu id-ul dat",
                "operationId": "9964293bc3eb922172f89f144436de11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idClient": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "$CNP": {
                                        "description": "",
                                        "type": "string",
                                        "example": "2821205341248"
                                    },
                                    "$nume": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Popescu"
                                    },
                                    "$prenume": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Maria"
                                    },
                                    "$nrTelefon": {
                                        "description": "",
                                        "type": "string",
                                        "example": "+40744055645"
                                    },
                                    "$email": {
                                        "description": "",
                                        "type": "string",
                                        "example": "popescu_maria01@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clienti"
                ],
                "summary": "Adauga clientul",
                "description": "Adauga clientul",
                "operationId": "3739227e52437c5676011fcc46a3f6d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "$CNP": {
                                        "description": "",
                                        "type": "string",
                                        "example": "2821205341248"
                                    },
                                    "$nume": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Popescu"
                                    },
                                    "$prenume": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Maria"
                                    },
                                    "$nrTelefon": {
                                        "description": "",
                                        "type": "string",
                                        "example": "+40744055645"
                                    },
                                    "$email": {
                                        "description": "",
                                        "type": "string",
                                        "example": "popescu_maria01@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/clienti/{id}": {
            "get": {
                "tags": [
                    "Clienti"
                ],
                "summary": "Returneaza clientul cu id-ul dat",
                "description": "Returneaza clientul cu id-ul dat",
                "operationId": "781265d2e3333ddf6dd01fcd8f78a3b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul clientului dorit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clienti"
                ],
                "summary": "Sterge clientul cu id-ul dat",
                "description": "Sterge clientul cu id-ul dat",
                "operationId": "e80705797027dc8808498bacc0a65aa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul clientului dorit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}